// src/screens/report/types.ts
export interface ReportType {
  id: string;
  label: string;
  description: string;
}